update roll-op to work with op stack v1.3.0
The major change is that the L1 genesis now includes the L2 contracts in the
genesis itself.

This requires retooling the deploy logic to handle two scenarios: either deploy
to a temporary geth, or to a real L1.

The deploy logic was moved to `l2_deploy.py`.

The deploy config generation logic, which is now also necessary for generating
the L1 genesis, was moved to `deploy_config.py`.